% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{add_metadata}
\alias{add_metadata}
\title{Utility function to add metadata to dataframe}
\usage{
add_metadata(data, metadata)
}
\arguments{
\item{data}{Data.frame to add the metadata to.}

\item{metadata}{Named list containing column name and value to assign.}
}
\value{
Data.frame containing the metadata.
}
\description{
Utility function to add metadata to dataframe
}
