% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_window.R
\name{average}
\alias{average}
\title{Average observations within a moving window}
\usage{
average(data)
}
\arguments{
\item{data}{Dataframe that contains the columns `time`, `id`, `x`, and `y`.}
}
\value{
Single-row dataframe with transformed columns `x` and `y`
}
\description{
Average observations within a moving window
}
