% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{bias}
\alias{bias}
\title{Compute Bias}
\usage{
bias(x)
}
\arguments{
\item{x}{Numeric vector of values to for which to compute the statistic.}
}
\value{
Numeric denoting the mean difference between the two vectors
}
\description{
Compute Bias
}
