% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_distribution.R
\name{compare_distribution}
\alias{compare_distribution}
\title{Compare two distributions}
\usage{
compare_distribution(x, y, alpha = 0.05)
}
\arguments{
\item{x}{Numeric vector of test values.}

\item{y}{Numeric vector of reference values.}

\item{alpha}{Numeric denoting the signficance level.}
}
\value{
Single-rowed data.frame containing the separate confidence intervals 
for \code{x} and \code{y}, as well as the confidence interval for the difference 
\code{x - y} and a logical denoting whether 0 is contained within the latter
interval.
}
\description{
This function will check whether two distributions differ significantly. Done
by first taking the difference in the distribution of the values provided in
\code{x} and \code{y}, then defining the (1 - \code{alpha}) * 100% confidence
interval of this distributions, and finally checking whether 0 is a part of 
that interval.
}
\details{
If 0 is part of this interval, the two distributions are said to not differ 
significantly, and vice-versa if 0 is not a part of this interval. 

This function therefore allows an analysis analogous to a posterior check, or 
otherwise absolute goodness-of-fit with a parametric bootstrap procedure.
}
