% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_position.R
\name{denormalize_position}
\alias{denormalize_position}
\title{Transform data from -1 and 1 to normal range}
\usage{
denormalize_position(x, min_x = NULL, max_x = NULL)
}
\arguments{
\item{x}{Vector of numerical data to be transformed back to their normal range}

\item{min_x}{Numeric denoting the minimal value that the data originally 
could take. Has no default and should therefore be provided in order for the 
transformation to happen.}

\item{max_x}{Numeric denoting the maximal value that the data originally 
could take. Has no default and should therefore be provided in order for the 
transformation to happen.}
}
\description{
Transform the delivered data from the -1 and 1 range back to their original 
range, based on the provided minimal and maximal value that the data can take.
}
