% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_filter.R
\name{dynamic_filter}
\alias{dynamic_filter}
\title{Use a dynamical model to filter data}
\usage{
dynamic_filter(data, parameters = NULL, ...)
}
\arguments{
\item{data}{Dataframe that contains the columns `time`, `id`, `x`, and `y`.}

\item{parameters}{List containing user-specified parameters. Defaults to 
the parameters estimated on the stationary calibration data.}

\item{...}{Arguments passed on to DEoptim.control}
}
\value{
Smoothed dataframe with a similar structure as `data`
}
\description{
This function uses VAR(1) parameters to estimate the true positions of the 
data. It does so through means of min-log-likelihood estimation where:
}
\details{
\bm{\mu}_t = \bm{y}_t - B \bm{y}_{t - 1} - \bm{\epsilon}_t
\bm{epsilon}_t \sim N(\bm{0}, \Sigma)

In this model, the values of $\bm{\mu}$ are estimates using the data and the 
"known" values of $\bm{y}$, $B$ and $\Sigma$. The values of the parameters 
$B$ and $\Sigma$ are estimated on the stationary data. 

Another approach to this kind of filtering might be more individualistic, 
requiring us to do a stationary calibration before each experiment and, 
ultimately, allowing us to estimate parameters for each tag for each session
separately. Something to be discussed, depending on the efficacy of this 
method in the synthetic simulation study.
}
