% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{kalman_filter}
\alias{kalman_filter}
\title{Smooth using a Kalman filter}
\usage{
kalman_filter(
  data,
  reverse = TRUE,
  internal = FALSE,
  model = "constant_velocity",
  .by = NULL,
  assumed_variance = 0.031^2,
  check = FALSE
)
}
\arguments{
\item{data}{Dataframe that contains the columns `time`, `id`, `x`, and `y`.}

\item{model}{String denoting which model to use.}

\item{.by}{String denoting whether the moving window should be taken with 
respect to a given grouping variable. Defaults to `NULL`.}
}
\value{
Smoothed dataframe with a similar structure as `data`
}
\description{
This is a higher-level function that will first look at whether the data needs
to be processed by a given variable. Then it will run the Kalman filter on 
these grouped data.
}
