% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{kalman_filter_individual}
\alias{kalman_filter_individual}
\title{Use Kalman filter on the data}
\usage{
kalman_filter_individual(
  data,
  reverse = TRUE,
  model = "constant_velocity",
  assumed_variance = 0.031^2
)
}
\arguments{
\item{data}{Dataframe that contains the columns `time`, `id`, `x`, and `y`.}

\item{model}{String denoting which of the models to use.}
}
\value{
Smoothed dataframe with the same structure as `data`
}
\description{
Use Kalman filter on the data
}
