% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{kf_innovation}
\alias{kf_innovation}
\title{Innovation step in the Kalman filter}
\usage{
kf_innovation(y, x, H, V, F)
}
\arguments{
\item{y}{Vector of measured values at time t + 1}

\item{x}{Vector of predicted values of the movement equation at time t + 1.}

\item{H}{Matrix relating the values in x to the values in y.}

\item{V}{Cholesky decomposition of the measurement noise covariance matrix.}

\item{F}{Cholesky decomposition of the estimation covariance matrix at time t + 1.}
}
\value{
Smoothed dataframe with a similar structure as `data`
}
\description{
Assumptions in the model:
 a) Mean of the process noise is equal to 0
}
