% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{kf_predict}
\alias{kf_predict}
\title{Predict step in the Kalman filter}
\usage{
kf_predict(x, A, u, B, W, F)
}
\arguments{
\item{x}{Vector of values of the movement equation at time t.}

\item{A}{Transition matrix, relating values of X at time t to those at t.}

\item{u}{Vector of values for the external variables at time t.}

\item{B}{Matrix that connects the external variables in u to the values in x.}

\item{W}{Cholesky decomposition of the process noise covariance matrix.}

\item{F}{Cholesky decomposition of the estimation covariance matrix at time t.}
}
\value{
List of predicted values for x and F at time t + 1
}
\description{
Assumptions in the model:
 a) Mean of the process noise is equal to 0
}
