% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{kf_update}
\alias{kf_update}
\title{Update step in the Kalman filter}
\usage{
kf_update(x, z, H, V, F, K)
}
\arguments{
\item{x}{Vector of predicted values of the movement equation at time t + 1.}

\item{z}{Vector of innovations at time t + 1}

\item{H}{Matrix relating the values in x to the values in y.}

\item{V}{Cholesky decomposition of the measurement noise covariance matrix.}

\item{F}{Cholesky decomposition of the estimation covariance matrix at time t + 1.}

\item{K}{Kalman gain at this updating step}
}
\value{
Smoothed dataframe with a similar structure as `data`
}
\description{
Update step in the Kalman filter
}
