% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{linear}
\alias{linear}
\title{Approximate positions with linear regression}
\usage{
linear(data, cols = NULL)
}
\arguments{
\item{data}{Dataframe that contains the columns `time`, `id`, `x`, and `y`.
Additionally, the variable `"index"` should be provided (automatically the 
case when using `moving_window`), as this will tell us which of the 
observations to use for our predictions.}
}
\value{
Single-row dataframe with transformed columns `x` and `y`
}
\description{
Estimate a linear regression in both the x- and the y-direction (i.e., 
with y ~ x and x ~ y) and take the best fitting one. We then approximate the 
real position by filling out the observed x and y in the equation and 
averaging observed (x, y) coordinates with the predicted ones. This should 
(hopefully) decrease the error in both dimensions instead of keeping the 
error in one dimension to predict the location in the other.
}
