% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load the experimental data}
\usage{
load_data()
}
\description{
This is a basic function that will load in the experimental data that we have
in a format that we can analyze. More specifically, it will combine the data
from files 'experiments.csv' -- which contains information on the experiments 
that were run -- and 'datapoints.csv' -- which contains the positions that 
were measured during each experiment.
}
