% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{mae}
\alias{mae}
\title{Compute MAE}
\usage{
mae(x)
}
\arguments{
\item{x}{Numeric vector of values to for which to compute the statistic.}
}
\value{
Numeric denoting the MAE of the difference between the two vectors
}
\description{
Compute MAE
}
