% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{middle}
\alias{middle}
\title{Take the middle observation within a moving window}
\usage{
middle(data, cols = NULL)
}
\arguments{
\item{data}{Dataframe that contains the columns `time`, `id`, `x`, and `y`.}
}
\value{
Single-row dataframe with transformed columns `x` and `y`
}
\description{
Take the middle observation within a moving window
}
