% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_window.R
\name{moving_window}
\alias{moving_window}
\title{Execute function on a moving window}
\usage{
moving_window(data, span, fx, .by = NULL)
}
\arguments{
\item{data}{Dataframe that contains the columns `time`, `id`, `x`, and `y`.}

\item{span}{Integer denoting the number of neighbours in each direction to 
take into account within each moving window.}

\item{fx}{Function to execute on the data within the moving window. Should 
take in a similar data structure to `data` and output a single-row dataframe
with columns `x` and `y` (representing the smoothed values of both variables).}

\item{.by}{String denoting whether the moving window should be taken with 
respect to a given grouping variable. Defaults to `NULL`}
}
\value{
Smoothed dataframe with a similar structure as `data`
}
\description{
Create a moving window of a given span and execute the specified function on 
the data within that window. In this project, a moving window is used to 
smooth the data.
}
\details{
Importantly, the span is a relative variable, that is it represents the 
number of neighbours to take in to account in each direction relative to the 
central data point to be smoothed. In other words, if you take `span = 2`, 
then you will smooth over the indices `c(-2, -1, 0, 1, 2)` where index `0` 
represents the datapoint to be smoothed.
}
