% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{name_plot}
\alias{name_plot}
\title{Create a pane with a chosen title in it}
\usage{
name_plot(x, ...)
}
\arguments{
\item{x}{Title to be plotted in an empty plot.}

\item{...}{Additional arguments for \code{geom_text}.}
}
\value{
Ggplot object
}
\description{
Create a pane with a chosen title in it
}
