% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_position.R
\name{normalize_position}
\alias{normalize_position}
\title{Transform data to -1 and 1}
\usage{
normalize_position(x, min_x = NULL, max_x = NULL)
}
\arguments{
\item{x}{Vector of numerical data to be transformed}

\item{min_x}{Numeric denoting the minimal value that the data `x` can take.
Defaults to the minimal observed value in `x`.}

\item{max_x}{Numeric denoting the maximal value that the data `x` can take.
Defaults to the maximal observed value in `x`.}
}
\description{
Transform the delivered data to the -1 and 1 range based on the provided 
minimal and maximal value that the data can take.
}
