% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot moving positions}
\usage{
plot(x, as_points = FALSE, per_iteration = TRUE, limits = NULL, ...)
}
\arguments{
\item{x}{Dataframe that contains the columns `time`, `id`, `x`, and `y`.}

\item{as_points}{Logical denoting whether to plot the positions as a single 
point (`TRUE`) or as a circle with a given orientation (see also `predped`, 
`FALSE`). Defaults to `FALSE`.}

\item{per_iteration}{Logical denoting whether to make multiple plots of 
positions at each time point (`TRUE`) or to just plot all measured locations 
at once (`FALSE`). If `FALSE`, `as_points` is ignored (will automatically be
`TRUE`). Defaults to `TRUE`}

\item{...}{Additional arguments to pass to ggplot2.}
}
\value{
List of plots, one for each time points in the dataframe.
}
\description{
Plot the pedestrian's movements based on the positions in a dataframe. This 
dataframe should consist of several columns, namely (a) a time variable that
orders the positions (`time`), (b) a grouping variable that denotes which 
positions belong to which pedestrian (`id`), and (c) the x and y positions 
themselves (`x` and `y` resp.).
}
\details{
Disclaimer: For this function to work properly, the time-variable should be 
an increasing number: It should not be in datatime format.
}
