% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_distortion.R
\name{polynomial_distortion}
\alias{polynomial_distortion}
\title{Correct systematic error through polynomial}
\usage{
polynomial_distortion(data, params = NULL, n = 1, simple = TRUE)
}
\arguments{
\item{data}{Dataframe containing the measured locations of x and y.}

\item{params}{Parameters as outputted by the least-squares procedure (k x 2, 
where k is the number of parameters)}

\item{n}{Degree of the polynomial. Defaults to `1`}

\item{simple}{Whether to only use main effects of `x` and `y` (`TRUE`) or to 
also include interaction effects of varying degrees (`FALSE`). Defaults to 
`TRUE`.}
}
\value{
Dataframe with corrected values of x and y
}
\description{
Use a polynomial of the n^th degree with the provided parameters to correct 
the systematric error in the measured locations `x` and `y`.
}
