% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{rmse}
\alias{rmse}
\title{Compute RMSE}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{Numeric vector of values to be compared to \code{y}.}

\item{y}{Numeric vector of values to be compared to \code{x}.}
}
\value{
Numeric denoting the RMSE of the difference between the two vectors
}
\description{
Compute RMSE
}
