% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summary_statistics}
\alias{summary_statistics}
\title{Compute summary statistics}
\usage{
summary_statistics(data, fx, .vars, .by = "id")
}
\arguments{
\item{data}{Data.frame containing the columns in \code{.vars} and for which 
to compute the summary statistics.}

\item{fx}{List of functions that compute the summary statistics of interest.
Should take in two numeric vectors. If named, the columns will be named after
the names of the list. Otherwise will be assigned another unique name.}

\item{.vars}{Character vector or matrix containing the columns for which to 
compute the summary statistics. If matrix, each row should contain the columns
to compare.}

\item{.by}{Character vector denoting the columns to group by. Defaults to
\code{"id"}.}
}
\value{
Data.frame containing the summary statistics of interest.
}
\description{
Overhead function that takes in the data and then computes all summary 
statistics contained in \code{fx}.
}
