% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{weighted_average}
\alias{weighted_average}
\title{Average observations within a moving window}
\usage{
weighted_average(
  data,
  weights = function(x) dnorm(x, mean = 0, sd = 1),
  .by = "index"
)
}
\arguments{
\item{data}{Dataframe that contains the columns `time`, `id`, `x`, and `y`.}

\item{weights}{Either a numeric vector of weights or a function that creates 
the weights to be assigned to each of the data points. Defaults to 
`\(x) dnorm(x, mean = 0, sd = 1)`}

\item{.by}{String denoting the column by which the weights should be determined
(i.e., of which the values are passed on to `weights`). Ignored if a numeric 
is provided for `weights`. Defaults to `"index"`.}
}
\value{
Single-row dataframe with transformed columns `x` and `y`
}
\description{
Average observations within a moving window
}
