% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{local_regression}
\alias{local_regression}
\title{Filter data through LOESS}
\usage{
local_regression(data, .by = NULL, degree = 1, ...)
}
\arguments{
\item{data}{Dataframe that contains the columns `time`, `id`, `x`, and `y`.}

\item{degree}{Integer denoting the degree of the polynomial.}
}
\value{
Single-row dataframe with transformed columns `x` and `y`
}
\description{
Uses the \code{loess} function in R to filter the data. Is a generalization 
and improvement of \code{linear} and \code{parabola} used with the 
\code{moving_window}.
}
