% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_efficiency.R
\name{pipeline_efficiency}
\alias{pipeline_efficiency}
\title{Compare preprocessed with actual positions}
\usage{
pipeline_efficiency(
  data,
  fx,
  .by = NULL,
  summary = list(bias = bias, rmse = rmse, mae = mae),
  .vars = c("X", "Y", "Dist", "x", "y", "dist"),
  summary.by = "id",
  path = file.path(".", "results"),
  filename = "",
  metadata = list(),
  n_cores = 1
)
}
\arguments{
\item{data}{Dataframe to be preprocessed}

\item{fx}{List of preprocessing pipelines to be executed. Each pipeline 
consists of another list with functiosn to use when preprocessing the data.
If names are added to this list, the names will be included in a separate 
column \code{function} in the resulting dataframes. Allows one to know which
data were created by which function. Otherwise includes number of the pipeline.}

\item{.by}{Character vector containing columns to group the preprocessing by. 
This only concerns columns at the highest level and is not passed on to the 
lower-level preprocessing functions. Defaults to \code{NULL}.}

\item{summary}{List of functions that will be used to summarize the results of
preprocessing. Defaults to computation of the RMSE, MAE, and mean difference.}

\item{.vars}{Character vector or matrix containing the columns for which to 
compute the summary statistics. If matrix, each row should contain the columns
to compare. Defaults to x, y, and distance.}

\item{summary.by}{Character vector containing variables to group by when 
creating the summary statistics. Defaults to \code{"id"}.}

\item{path}{Path under which to save the datafiles that result from this 
function. Defaults to a "results" folder at the current directory.}

\item{filename}{Name to use as an identifier for the results of this analysis.
Defaults to an empty string.}

\item{metadata}{Named list containing metadata that should be added to the 
preprocessed datafiles. Consists of column name and value to assign to it.
Defaults to an empty list.}

\item{n_cores}{Integer denoting how many cores to use for the analysis. 
Defaults to \code{1}.}
}
\value{
Returns \code{NULL} and instead saves several datafiles under the path 
specified by the user
}
\description{
This function preprocesses the provided data and then checks how close it gets
to the actual data. Importantly, this process assumes that you have the actual 
positions under columns "x_actual" and "y_actual" while the measured 
positions are under "x" and "y". Preprocessed measures will also end up under
"x" and "y", but only after preprocessing.
}
\details{
Depends on the function \code{\link[nameless]{summary_statistics}} which 
uses measured/filtered and actual positions to compute mean deviation, 
absolute mean deviation, and root mean squared error.
}
