% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{rmse}
\alias{rmse}
\title{Compute RMSE}
\usage{
rmse(x)
}
\arguments{
\item{x}{Numeric vector of values to for which to compute the statistic.}
}
\value{
Numeric denoting the RMSE of the difference between the two vectors
}
\description{
Compute RMSE
}
